+++ src/libmach/coff32/coff32.h
@@ -39,5 +39,5 @@ extern int coff32xsetidx(int order,
 extern int coff32xgetidx(int order,
                          long *nsyms, char ***namep, long **offsp, FILE *fp);
 
-extern Symbol *coff32getsym(Obj *obj, long *idx, Symbol *sym);
-extern Section *coff32getsec(Obj *obj, long *idx, Section *sec);
+extern Symbol *coff32getsym(Obj *obj, int *idx, Symbol *sym);
+extern Section *coff32getsec(Obj *obj, int *idx, Section *sec);
