+++ src/libmach/coff32/coff32getsym.c
@@ -54,7 +54,7 @@ symname(Coff32 *coff, SYMENT *ent)
 }
 
 Symbol *
-coff32getsym(Obj *obj, long *idx, Symbol *sym)
+coff32getsym(Obj *obj, int *idx, Symbol *sym)
 {
 	long n = *idx;
 	SYMENT *ent;
